feat: 돈관련 자료형 BigInteger, BigDecimal 변경
int 자료형의 최대 값이 2,147,483,647라서 돈 액수를 표현하기에 부족하다고 판단되어 자료형을 변경합니다.

+++ src/main/java/lotto/domain/Cashier.java
@@ -1,10 +1,11 @@
 package lotto.domain;
 
+import java.math.BigInteger;
 import java.util.Objects;
 
 public class Cashier {
 
-    public static final int LOTTO_PURCHASING_UNIT = 1000;
+    public static final BigInteger LOTTO_PURCHASING_UNIT = BigInteger.valueOf(1000);
 
     private final Money money;
 
@@ -17,8 +18,10 @@ public class Cashier {
         this(new Money(money));
     }
 
+    // TODO: 메소드 이름 바꾸기
     public int findCountOfPlayLotto() {
-        return this.money.divideBy(LOTTO_PURCHASING_UNIT);
+        return this.money.divideBy(LOTTO_PURCHASING_UNIT)
+                .intValue();
     }
 
     private void validateThousandUnits(Money money) {
